feat: Create Meal Remote Feature
