+++ feature/meal/src/commonMain/kotlin/com/b1nd/dodam/meal/model/MealUiState.kt
@@ -1,0 +1,13 @@
+package com.b1nd.dodam.meal.model
+
+import com.b1nd.dodam.data.meal.model.Meal
+import kotlinx.collections.immutable.ImmutableList
+import kotlinx.datetime.LocalDate
+
+sealed interface MealUiState {
+    data class Success(
+        val meals: ImmutableList<Meal>
+    ) : MealUiState
+    data object Error : MealUiState
+    data object Loading : MealUiState
+}
