+++ feature/meal/src/commonMain/kotlin/com/b1nd/dodam/meal/viewmodel/MealViewModel.kt
@@ -6,6 +6,7 @@ import com.b1nd.dodam.common.date.DodamDate
 import com.b1nd.dodam.common.result.Result
 import com.b1nd.dodam.data.meal.MealRepository
 import com.b1nd.dodam.data.meal.model.Meal
+import com.b1nd.dodam.meal.model.MealUiState
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
@@ -18,48 +19,28 @@ import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 
 class MealViewModel : ViewModel(), KoinComponent {
-    private val current = DodamDate.now()
-    private var getOnlyOnce = true
 
-    private var meals: PersistentList<Meal> = persistentListOf()
+    private var loadOnce: Pair<Int, Int>? = null
     private val mealRepository: MealRepository by inject()
-
-    init {
-        getMealOfMonth(current.year, current.monthNumber)
-        getOnlyOnce = true
-    }
 
     private val _mealUiState: MutableStateFlow<MealUiState> = MutableStateFlow(MealUiState.Loading)
     val mealUiState: StateFlow<MealUiState> = _mealUiState.asStateFlow()
 
     fun getMealOfMonth(year: Int, month: Int) {
-        if (getOnlyOnce) {
-            getOnlyOnce = false
-            mealRepository.getMealOfMonth(year, month)
-                .onEach { result ->
-                    when (result) {
-                        is Result.Success -> _mealUiState.emit(
-                            MealUiState.Success(
-                                meals.addAll(
-                                    result.data.filter {
-                                        it.date >= current.date
-                                    },
-                                ).also {
-                                    meals = it
-                                },
-                            ),
+        if (loadOnce?.first == year && loadOnce?.second == month) { return }
+        loadOnce = year to month
+        mealRepository.getMealOfMonth(year, month)
+            .onEach { result ->
+                when (result) {
+                    is Result.Success -> _mealUiState.emit(
+                        MealUiState.Success(
+                            meals = result.data,
                         )
+                    )
 
-                        is Result.Loading -> _mealUiState.emit(MealUiState.Loading)
-                        is Result.Error -> _mealUiState.emit(MealUiState.Error)
-                    }
-                }.launchIn(viewModelScope)
-        }
+                    is Result.Loading -> _mealUiState.emit(MealUiState.Loading)
+                    is Result.Error -> _mealUiState.emit(MealUiState.Error)
+                }
+            }.launchIn(viewModelScope)
     }
-}
-
-sealed interface MealUiState {
-    data class Success(val meals: ImmutableList<Meal>) : MealUiState
-    data object Error : MealUiState
-    data object Loading : MealUiState
 }
