+++ feature/meal/src/main/java/com/b1nd/dodam/meal/preview/MealScreenPreview.kt
@@ -1,50 +1,9 @@
 package com.b1nd.dodam.meal.preview
 
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.gestures.detectVerticalDragGestures
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.requiredHeight
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableFloatStateOf
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.geometry.CornerRadius
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.times
-import androidx.compose.ui.util.fastForEach
 import com.b1nd.dodam.designsystem.DodamTheme
-import com.b1nd.dodam.designsystem.animation.rememberBounceIndication
-import com.b1nd.dodam.designsystem.component.CalendarDate
-import com.b1nd.dodam.designsystem.component.DodamDatePickerDefaults
-import com.b1nd.dodam.designsystem.component.DodamDatePickerState
-import com.b1nd.dodam.designsystem.component.rememberDodamDatePickerState
-import com.b1nd.dodam.logging.KmLogging
-import com.b1nd.dodam.meal.ExpandableCalendar
-import com.b1nd.dodam.meal.NewMealScreen
+import com.b1nd.dodam.meal.MealScreen
 
 @Preview
 @Composable
@@ -52,6 +11,6 @@ fun MealScreenPreview(
 
 ) {
     DodamTheme {
-        NewMealScreen()
+        MealScreen()
     }
 }
