+++ ui/src/commonMain/composeResources/drawable/ic_colored_cookedrice.xml
@@ -1,0 +1,48 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M13.987,22.506H9.945C5.287,22.506 1.507,18.726 1.507,14.068V10.123H22.425V14.068C22.425,18.726 18.645,22.506 13.987,22.506Z"
+      android:fillColor="#F8312F"/>
+  <path
+      android:pathData="M16.44,5.796C18.825,5.796 20.767,7.738 20.767,10.123C20.767,12.508 18.825,14.451 16.44,14.451H7.485C5.1,14.451 3.158,12.508 3.158,10.123C3.158,7.738 5.1,5.796 7.485,5.796H16.44ZM16.44,4.138H7.485C4.177,4.138 1.5,6.816 1.5,10.123C1.5,13.431 4.177,16.108 7.485,16.108H16.433C19.74,16.108 22.417,13.431 22.417,10.123C22.425,6.823 19.74,4.138 16.44,4.138Z"
+      android:fillColor="#CA0B4A"/>
+  <path
+      android:pathData="M3.176,9.212L3.162,9.085C3.147,8.941 3.123,8.793 3.096,8.626L3.088,8.576C3.022,8.171 2.931,7.615 3.084,7.076C3.112,6.974 3.136,6.883 3.158,6.801C3.258,6.427 3.309,6.239 3.419,6.129L3.418,6.123C3.528,5.982 3.619,5.856 3.693,5.728C3.811,5.523 3.905,5.285 4.022,4.982L4.042,4.931C4.211,4.494 4.443,3.892 4.9,3.435C5.35,2.986 5.94,2.754 6.368,2.585L6.418,2.566C6.727,2.444 6.965,2.349 7.167,2.228L7.172,2.224L7.177,2.221C7.373,2.109 7.566,1.954 7.816,1.752L7.858,1.718C8.223,1.421 8.731,1.009 9.369,0.841C9.98,0.675 10.611,0.774 11.069,0.845L11.126,0.854C11.451,0.905 11.71,0.943 11.955,0.943C12.2,0.943 12.459,0.905 12.784,0.854L12.84,0.845C13.299,0.774 13.929,0.675 14.541,0.841C15.179,1.009 15.686,1.421 16.052,1.718L16.094,1.752C16.344,1.954 16.537,2.108 16.732,2.221C16.937,2.339 17.175,2.433 17.478,2.55L17.53,2.57C17.967,2.739 18.569,2.971 19.025,3.428C19.48,3.883 19.713,4.485 19.881,4.92L19.9,4.972C20.019,5.278 20.113,5.516 20.231,5.718L20.232,5.721C20.319,5.872 20.432,6.023 20.573,6.199L20.572,6.202C20.766,6.49 20.772,6.514 20.839,6.777C20.858,6.851 20.882,6.945 20.916,7.069C21.069,7.609 20.978,8.165 20.912,8.569L20.904,8.619C20.877,8.789 20.853,8.939 20.837,9.085L20.825,9.202C20.725,11.588 18.66,13.5 16.137,13.5H7.863C5.344,13.5 3.281,11.593 3.176,9.212Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M20.858,9.554C20.829,9.743 20.799,9.933 20.78,10.122H20.768V10.123C20.768,12.508 18.825,14.451 16.44,14.451H7.485C5.1,14.451 3.158,12.508 3.158,10.123V10.122H3.144C3.125,9.935 3.096,9.748 3.067,9.562C2.99,9.062 2.913,8.569 3.038,8.113C3.163,7.633 3.487,7.234 3.807,6.838C4.001,6.598 4.195,6.359 4.343,6.103C4.495,5.839 4.608,5.546 4.722,5.253C4.905,4.78 5.087,4.309 5.43,3.966C5.768,3.628 6.231,3.446 6.693,3.264C6.991,3.146 7.288,3.029 7.552,2.871C7.808,2.723 8.047,2.53 8.288,2.335C8.683,2.015 9.082,1.692 9.563,1.566C10.018,1.441 10.511,1.518 11.011,1.595C11.325,1.644 11.642,1.693 11.955,1.693C12.268,1.693 12.585,1.644 12.899,1.595C13.399,1.518 13.892,1.441 14.347,1.566C14.828,1.692 15.227,2.015 15.623,2.335C15.863,2.53 16.102,2.723 16.358,2.871C16.621,3.023 16.915,3.136 17.208,3.25C17.681,3.433 18.152,3.615 18.495,3.958C18.836,4.299 19.019,4.771 19.201,5.243C19.315,5.538 19.429,5.833 19.583,6.096C19.73,6.351 19.924,6.59 20.118,6.831C20.438,7.226 20.762,7.625 20.888,8.106C21.012,8.562 20.935,9.054 20.858,9.554Z"
+      android:fillColor="#F4F4F4"/>
+  <path
+      android:pathData="M11.288,3.358L12,4.588C12.158,4.866 12.068,5.211 11.79,5.368C11.512,5.526 11.167,5.428 11.01,5.158L10.297,3.928C10.14,3.651 10.238,3.306 10.507,3.148C10.785,2.991 11.13,3.088 11.288,3.358Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M14.55,5.946L14.813,4.551C14.873,4.243 15.172,4.041 15.48,4.093C15.795,4.153 15.998,4.453 15.938,4.761L15.675,6.156C15.615,6.471 15.315,6.673 15.007,6.613C14.693,6.553 14.49,6.253 14.55,5.946Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M9.69,9.478L9.113,8.181C8.985,7.888 9.113,7.558 9.398,7.431C9.683,7.303 10.02,7.431 10.148,7.716L10.725,9.013C10.852,9.298 10.725,9.636 10.44,9.763C10.155,9.891 9.818,9.763 9.69,9.478Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M6.922,12.748L7.185,11.353C7.237,11.038 7.537,10.836 7.852,10.896C8.167,10.956 8.37,11.256 8.31,11.563L8.047,12.958C7.987,13.273 7.687,13.476 7.38,13.416C7.065,13.356 6.862,13.056 6.922,12.748Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M5.865,8.593L4.605,9.253C4.327,9.396 4.215,9.741 4.365,10.018C4.507,10.296 4.852,10.408 5.13,10.258L6.39,9.598C6.667,9.456 6.78,9.111 6.63,8.833C6.487,8.556 6.142,8.451 5.865,8.593Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M6.78,4.776L5.932,5.916C5.745,6.171 5.797,6.523 6.052,6.711C6.307,6.898 6.66,6.846 6.847,6.591L7.695,5.451C7.882,5.196 7.83,4.843 7.575,4.656C7.327,4.468 6.967,4.521 6.78,4.776Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M11.198,12.426L12.045,11.286C12.233,11.038 12.585,10.986 12.84,11.166C13.095,11.353 13.148,11.706 12.96,11.961L12.113,13.101C11.925,13.356 11.573,13.408 11.318,13.221C11.063,13.033 11.01,12.681 11.198,12.426Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M17.64,8.256L18.563,9.336C18.765,9.576 19.125,9.606 19.365,9.403C19.605,9.201 19.635,8.841 19.433,8.601L18.51,7.521C18.308,7.281 17.948,7.251 17.708,7.453C17.468,7.663 17.438,8.016 17.64,8.256Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M18.03,12.411L17.317,11.181C17.16,10.911 16.815,10.813 16.538,10.971C16.267,11.128 16.17,11.473 16.327,11.751L17.04,12.981C17.198,13.251 17.542,13.348 17.82,13.191C18.09,13.033 18.188,12.688 18.03,12.411Z"
+      android:fillColor="#BEBEBE"/>
+  <path
+      android:pathData="M13.86,7.911L15.165,8.473C15.45,8.593 15.585,8.931 15.465,9.223C15.337,9.516 15,9.651 14.715,9.523L13.41,8.961C13.118,8.833 12.983,8.496 13.11,8.211C13.238,7.918 13.575,7.783 13.86,7.911Z"
+      android:fillColor="#BEBEBE"/>
+</vector>
